Revert "Add Express Server and integrate with MongoDB Atlas"
