+++ index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="id">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Selamat Datang di Tes Psikologi Entrepreneur</title>
+    <script src="https://cdn.tailwindcss.com"></script>
+</head>
+<body class="bg-gray-900 justify-center flex">
+    <div class="max-w-lg text-center mt-10 border-8 m-8 bg-gray-200 p-5 rounded-[40px] shadow-lg">
+        <h1 class="text-3xl lg:text-4xl font-bold mb-4 text-gray-900 mt-10">SELAMAT DATANG!</h1>
+        <h1 class="text-3xl lg:text-4xl font-bold mb-4 text-gray-900 ">DI TEST PSIKOLOGI ENTERPRENEUR</h1>
+        <p class="text-lg lg:text-ms mb-6 text-gray-900 mt-10">KREKI ( KOMUNITAS RELAWAN EMERGENSI KESEHATAN INDONESIA ) bersama INDOHCF ( INDONESIA HEALTHCARE FORUM ) , mengajak Bpk /Ibu / Sdr / Sdri melakukan self assesment , apakah kita mempunyai jiwa entrepreneur  
+Sekaligus memotivasi agar jiwa *entrepreneur*kita meningkat dg berusaha  meraih score tinggi .
+</p>
+        <p class="text-lg lg:text-ms mb-6 text-gray-900 mt-1">Jawab setiap pertanyaan dengan jujur dan hasilnya akan membantu Anda memahami diri Anda lebih baik.</p>
+        <a href="tes.html" class="mt-10 flex justify-center gap-2 items-center 
+        mx-auto shadow-xl text-lg bg-gray-50 backdrop-blur-md lg:font-semibold 
+        isolation-auto border-gray-50 before:absolute before:w-full before:transition-all 
+        before:duration-700 before:hover:w-full before:-left-full before:hover:left-0 
+        before:rounded-full before:bg-gray-900 hover:text-gray-50 before:-z-10 before:aspect-square 
+        before:hover:scale-150 before:hover:duration-700 relative z-10 px-8 py-4 overflow-hidden border-2 
+        rounded-full group">Mulai</a>
+    </div>
+</body>
+</html>
