--- package.json
@@ -1,22 +1,0 @@
-{
-  "name": "tes-psikologi-enterpreneur",
-  "version": "1.0.0",
-  "description": "buaat tugas bapak",
-  "main": "index.js",
-  "scripts": {
-    "start": "node -r dotenv/config ./src/server.js",
-    "dev": "nodemon -r dotenv/config ./src/server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "dotenv": "^16.4.5",
-    "nodemon": "^3.1.4"
-  },
-  "dependencies": {
-    "ejs": "^3.1.10",
-    "express": "^4.21.0",
-    "mongoose": "^8.6.2"
-  }
-}
